+++ test/libsolidity/syntaxTests/nameAndTypeResolution/212_uninitialized_mapping_array_variable_050.sol
@@ -1,9 +1,0 @@
-pragma experimental "v0.5.0";
-contract C {
-    function f() pure public {
-        mapping(uint => uint)[] storage x;
-        x;
-    }
-}
-// ----
-// DeclarationError: (82-115): Uninitialized storage pointer.